python_backend/requirements.txt
aiohappyeyeballs==2.4.2
aiohttp==3.10.7
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.7.0
anthropic==0.34.2
anyio==4.6.0
appnope==0.1.4
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
blinker==1.8.2
cachelib==0.13.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
debugpy==1.8.1
decorator==5.1.1
deprecation==2.1.0
distro==1.9.0
exceptiongroup==1.2.1
executing==2.0.1
fastapi==0.115.0
filelock==3.16.1
Flask==3.0.3
Flask-Cors==5.0.0
Flask-Session==0.8.0
Flask-WTF==1.2.1
frozenlist==1.4.1
fsspec==2024.9.0
gotrue==2.8.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.25.1
hyperframe==6.0.1
idna==3.10
ipykernel==6.29.4
ipython==8.24.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.5.0
jupyter_client==8.6.1
jupyter_core==5.7.2
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
msgspec==0.18.6
multidict==6.1.0
nest-asyncio==1.6.0
ngrok==1.4.0
openai==1.50.2
packaging==24.0
parso==0.8.4
pexpect==4.9.0
platformdirs==4.2.1
postgrest==0.16.11
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.17.2
PyJWT==2.9.0
PyQt5==5.15.10
PyQt5-Qt5==5.15.13
PyQt5-sip==12.13.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
pyzmq==26.0.2
realtime==2.0.4
requests==2.32.3
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.38.6
storage3==0.7.7
StrEnum==0.4.15
supabase==2.7.4
supafunc==0.5.1
timedelta==2020.12.3
tokenizers==0.20.0
tornado==6.4
tqdm==4.66.5
traitlets==5.14.3
twilio==9.3.2
typing_extensions==4.11.0
urllib3==2.2.3
wcwidth==0.2.13
websockets==13.1
Werkzeug==3.0.4
WTForms==3.1.2
yarl==1.13.1
